insert different naming in WebIf output of entitlements
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5524 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -464,7 +464,8 @@ typedef struct v_ban {					// failban listmember
 
 typedef struct s_entitlement {			// contains entitlement Info
 	uint16_t		id;					// the element ID
-	uint8_t			type;				// enumerator for tier,chid whatever
+	uint32_t		type;				// enumerator for tier,chid whatever
+										// 0="", 1="Package", 2="PPV-Event", 3="chid", 4="tier"
 	uint16_t		caid;				// the caid of element
 	uint32_t		provid;				// the provid of element
 	uint16_t		class;				// the class needed for some systems
